fix: option to resubmit missed transaction retry feature also added to retry resubmitting missed transaction
