+++ .env.template
@@ -32,6 +32,7 @@ TRANSACTIONHISTORY_PACS002_COLLECTION=transactionHistoryPacs002
 # ELASTIC APM
 APM_ACTIVE=true
 APM_SERVICE_NAME=ppa-batch-to-tms
+RETRY=3
 # APM_URL=127.0.0.1:4857
 APM_URL=http://apm:8200
 APM_SECRET_TOKEN=""
